# docs/implementation/contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AI Cloud API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /models/{model_id}/admission:
    parameters: [{$ref: '#/components/parameters/ModelID'}]
    get:
      operationId: getModelAdmission
      responses:
        '200': {description: Admission view, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: appendModelAdmissionEvidence
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdmissionEvidence'}}}}
      responses:
        '201': {description: Admission evidence, content: {application/json: {schema: {$ref: '#/components/schemas/AdmissionEvidence'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /tasks/{task_id}/model:
    parameters: [{$ref: '#/components/parameters/TaskID'}]
    post:
      operationId: executeTaskModel
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModelExecutionRequest'}}}}
      responses:
        '200': {description: Model result, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
        default: {$ref: '#/components/responses/Error'}
  /tasks/{task_id}/trace:
    parameters: [{$ref: '#/components/parameters/TaskID'}, {$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/PageToken'}]
    get:
      operationId: listTaskTrace
      responses:
        '200': {description: Trace evidence, content: {application/json: {schema: {$ref: '#/components/schemas/ListResponse'}}}}
        default: {$ref: '#/components/responses/Error'}
  /tasks/{task_id}/evaluations:
    parameters: [{$ref: '#/components/parameters/TaskID'}]
    get:
      operationId: listTaskEvaluations
      parameters: [{$ref: '#/components/parameters/PageSize'}, {$ref: '#/components/parameters/PageToken'}]
      responses:
        '200': {description: Evaluations, content: {application/json: {schema: {$ref: '#/components/schemas/ListResponse'}}}}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: createTaskEvaluation
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EvaluationRequest'}}}}
      responses:
        '201': {description: Evaluation created, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 256}
    PageSize:
      name: pageSize
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    PageToken:
      name: pageToken
      in: query
      schema: {type: string}
    TaskID:
      name: task_id
      in: path
      required: true
      schema: {type: string, minLength: 1}
    ModelID:
      name: model_id
      in: path
      required: true
      schema: {type: string, minLength: 1}
...
  responses:
    Error:
      description: Stable API error envelope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
...
    ListResponse:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {type: object, additionalProperties: true}}
        nextPageToken: {type: string}
...
    SchemaReference:
      type: object
      additionalProperties: false
      required: [name, version]
      properties:
        name: {type: string, minLength: 1}
        version: {type: string, minLength: 1}
    ModelExecutionRequest:
      type: object
      additionalProperties: false
      required: [taskType, instruction, outputSchema]
      properties:
        requestId: {type: string}
        taskType:
          type: string
          enum: [GeneratePlan, GeneratePatch, ExplainRisk, GenerateRollback, GenerateValidationReport, SummarizeState, RepairYAML, ExplainPolicyFailure]
        userId: {type: string}
        riskHint: {type: string}
        systemPrompt: {type: string}
        instruction: {type: string, minLength: 1}
        context: {type: object, additionalProperties: true}
        outputSchema: {$ref: '#/components/schemas/SchemaReference'}
        safetyPolicy: {$ref: '#/components/schemas/SchemaReference'}
        maxOutputTokens: {type: integer, minimum: 0}
        temperature: {type: number}
    EvaluationRequest:
      type: object
      additionalProperties: false
      required: [config, metrics, thresholds]
      properties:
        config: {type: object, additionalProperties: true}
        metrics: {type: object, additionalProperties: true}
        thresholds: {type: object, additionalProperties: true}
        rawOutput: {type: string}
...
    AdmissionEvidence:
      type: object
      additionalProperties: false
      properties:
        id: {type: string}
        modelId: {type: string}
        modelVersion: {type: string}
        status: {type: string, enum: [evidence-collected, reviewed, approved, restricted, rejected, revoked]}
        licenseId: {type: string}
        licenseTextRef: {type: string}
        sourceRef: {type: string}
        upstreamModels: {type: array, items: {type: string}}
        datasetRefs: {type: array, items: {type: string}}
        commercialUseAllowed: {type: boolean}
        hostedServiceAllowed: {type: boolean}
        redistributionAllowed: {type: boolean}
        noticeRequired: {type: boolean}
        noticeRef: {type: string}
        artifactDigest: {type: string}
        artifactSignature: {type: string}
        securityScanRef: {type: string}
        reviewer: {type: string}
        reviewedAt: {type: [string, 'null'], format: date-time}
        evidenceDigest: {type: string}
        createdAt: {type: string, format: date-time}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, trace_id, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}
            trace_id: {type: string}
            retryable: {type: boolean}
            details: {type: object, additionalProperties: true}